requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.5.0
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
